Commit messages are hard
